refactor: move max_reduce_except_dim() to utils.py
