+++ pytorch_optimizer/optimizer/sm3.py
@@ -3,6 +3,7 @@ from torch.optim.optimizer import Optimizer
 
 from pytorch_optimizer.base.optimizer import BaseOptimizer
 from pytorch_optimizer.base.types import CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.optimizer.utils import max_reduce_except_dim
 
 
 class SM3(Optimizer, BaseOptimizer):
@@ -32,7 +33,7 @@ class SM3(Optimizer, BaseOptimizer):
 
         self.validate_parameters()
 
-        defaults: DEFAULTS = {'lr': lr, 'momentum': momentum, 'beta': beta}
+        defaults: DEFAULTS = {'lr': lr, 'momentum': momentum, 'beta': beta, 'eps': eps}
         super().__init__(params, defaults)
 
     def validate_parameters(self):
@@ -52,21 +53,6 @@ class SM3(Optimizer, BaseOptimizer):
 
                 state['step'] = 0
                 state['momentum_buffer'] = torch.zeros_like(p)
-
-    @staticmethod
-    def max_reduce_except_dim(x: torch.Tensor, dim: int) -> torch.Tensor:
-        r"""Perform reduce-max along all dimensions except the given dim."""
-        rank: int = len(x.shape)
-        if rank == 0:
-            return x
-
-        if dim >= rank:
-            raise ValueError(f'[-] given dim is bigger than rank. {dim} >= {rank}')
-
-        for d in range(rank):
-            if d != dim:
-                x = x.max(dim=d, keepdim=True).values
-        return x
 
     @staticmethod
     def make_sparse(grad: torch.Tensor, values: torch.Tensor) -> torch.Tensor:
@@ -118,17 +104,14 @@ class SM3(Optimizer, BaseOptimizer):
                         update_values.mul_(beta)
                     update_values.addcmul_(grad._values(), grad._values(), value=1.0 - beta)
 
-                    nu_max = self.max_reduce_except_dim(
-                        x=self.make_sparse(grad, update_values).to_dense(),
-                        dim=0,
-                    ).squeeze_()
+                    nu_max = max_reduce_except_dim(self.make_sparse(grad, update_values).to_dense(), 0).squeeze_()
 
                     if beta > 0.0:
                         torch.max(acc, nu_max, out=acc)
                     else:
                         acc.copy_(nu_max)
 
-                    update_values.add_(self.eps).rsqrt_().mul_(grad._values())
+                    update_values.add_(group['eps']).rsqrt_().mul_(grad._values())
 
                     update = self.make_sparse(grad, update_values)
                 else:
@@ -142,13 +125,13 @@ class SM3(Optimizer, BaseOptimizer):
 
                     for i in range(rank):
                         acc = state[f'accumulator_{i}']
-                        nu_max = self.max_reduce_except_dim(update, i)
+                        nu_max = max_reduce_except_dim(update, i)
                         if beta > 0.0:
                             torch.max(acc, nu_max, out=acc)
                         else:
                             acc.copy_(nu_max)
 
-                    update.add_(self.eps).rsqrt_().mul_(grad)
+                    update.add_(group['eps']).rsqrt_().mul_(grad)
 
                     if momentum > 0.0:
                         m = state['momentum_buffer']
